fix: invoke compare
