--- packages/faas-cli-command-core/test/tmp/1.from
@@ -1,1 +1,0 @@
-1582625733792
