--- packages/faas-cli-command-core/test/tmp/1.to
@@ -1,1 +1,0 @@
-1582625733897
