--- packages/faas-cli-command-core/test/tmp/2.from
@@ -1,1 +1,0 @@
-1582625734002
