+++ packages/serverless-invoke/package.json
@@ -17,6 +17,7 @@
     "fs-extra": "^8.1.0",
     "ts-node": "^8.5.2",
     "urllib": "^2.34.1",
+    "globby": "^10.0.1",
     "websocket": "^1.0.30"
   },
   "devDependencies": {
